+++ Backend/src/main/java/org/ranking_app/service/user/UserDeleterService.java
@@ -1,5 +1,6 @@
 package org.ranking_app.service.user;
 
+import jakarta.servlet.http.HttpServletRequest;
 import org.ranking_app.model.user.User;
 import org.ranking_app.repository.user.JpaUserRepository;
 import org.springframework.stereotype.Service;
@@ -17,7 +18,13 @@ public class UserDeleterService {
         this.userFinderService = userFinderService;
     }
 
-    public void delete(Long id) {
+    public void delete(Long id, HttpServletRequest httpRequest) {
+        User authenticatedUser = userFinderService.findAuthenticatedUser(httpRequest);
+
+        if (!authenticatedUser.getAdmin()) {
+            throw new SecurityException("Acceso denegado");
+        }
+
         User user = userFinderService.find(id);
         jpaUserRepository.delete(user);
     }
